backend/requirements.txt
django==5.2.8
mysqlclient==1.1.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
celery==5.5.3
redis==7.1.0
faster-whisper==1.2.1
google-generativeai==0.8.5